contrib/verify: move helpers into their own file
This will make it simpler to extend the capabilities of the verifier in
the next commit without polluting too much the verify_test.go file.
