+++ Makefile
@@ -297,8 +297,8 @@ bpf-test:
 
 .PHONY: verify
 verify: tetragon-bpf ## Verify BPF programs.
-	$(GO) test contrib/verify/verify_test.go -run TestExtractKernelVersion
-	$(GO) test -exec 'sudo -E' contrib/verify/verify_test.go -v -run TestVerifyTetragonPrograms -tetragon-dir=$(CURDIR)/bpf/objs $(if $(DEBUG),-debug)
+	$(GO) test ./contrib/verify/ -skip TestVerifyTetragonPrograms
+	$(GO) test -exec 'sudo -E' ./contrib/verify/ -v -run TestVerifyTetragonPrograms -tetragon-dir=$(CURDIR)/bpf/objs $(if $(DEBUG),-debug)
 
 .PHONY: alignchecker
 alignchecker: ## Run alignchecker.
